TFC-12817 - Upgrade to Spring Security 4.0. Fix Unit Tests
Motivation for this was the patch here:
https://github.com/spring-projects/spring-flex/pull/10
But I also had to update the namespaces in the xml contexts in order to
get the unit tests working.

Also, needed to set the Thread Local on every Test in AMFViewTests, as I
assume later versions of JUnit use separate threads for each call.